Update README.md
Within this document, I have encapsulated a thorough description of the project, offering insight into its purpose, features, and the technology stack employed. Moreover, I have uploaded every file integral to the project, encompassing Java source code, HTML, CSS, and JavaScript files. To provide an immediate visual impression, the PowerPoint presentation file is included, offering a comprehensive overview of the website. This presentation not only serves as a visual guide but also contains multiple screenshots, allowing viewers to get a firsthand look at the project's interface and functionality. Feel free to explore the files to gain a deeper understanding of the intricacies of the project and its user-centric design.